feat(loan): add loan creation service with validation
- Implement loan creation endpoint with OTR, tenor, and asset validation
- Add database models and repository methods for loan operations
- Include loan service in main application router
- Update database schema with installment_amount field

+++ apps/models/customer_tenor.go
@@ -8,7 +8,7 @@ type CustomerTenors struct {
 	ID              string    `gorm:"type:uuid;default:uuid_generate_v4();primaryKey"`
 	CustomerID      string    `gorm:"type:uuid;column:customer_id"`
 	TenorID         string    `gorm:"type:uuid;column:tenor_id"`
-	LimitLoanAmount int       `gorm:"column:limit_loan_amount"`
+	LimitLoanAmount float64   `gorm:"column:limit_loan_amount"`
 	CreatedAt       time.Time `gorm:"column:created_at"`
 	UpdatedAt       time.Time `gorm:"column:updated_at"`
